requirements.txt
python==3.7.6
numpy==1.18.5
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-model-optimization==0.4.1
tensorflow-probability==0.9.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-layer-normalization==0.14.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.42.0
opencv-contrib-python==4.1.1.26
opencv-python==4.2.0.32
tqdm==4.42.1
imgaug==0.4.0
Pillow==7.0.0
imagecorruptions==1.1.0